requirements.txt
Django>=5.0,<6.0
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0
Pillow>=10.0.0
cryptography>=41.0.0
redis>=5.0.0
django-cors-headers>=4.0.0
whitenoise>=6.6.0
python-decouple>=3.8
aiofiles>=23.0.0
